restructure project and add Makefile
Restructures the project by moving osquery-related code to cmd/osquery directory
and adds a Makefile for better daemon management. The changes include:

- Add Makefile with osqueryd daemon management commands
- Move core application logic to separate app.go file
- Restructure main.go to use the new Application type
- Move all osquery related files to cmd/osquery directory
