+++ cmd/osquery/LICENSE
