+++ cmd/osquery/README.md
