+++ cmd/osquery/run_osquery_temp.sh
